Ensure deviatoric stress.
Most of FEvoR's computations work with either the Cauchy stress or the
deviatoric stress tensor. However, the polygonization ratio and the
determination of the dominate stress state in the recrystallization
calculates *do* assume a deviatoric stress.

This commit ensures that the stress applied to a FEvoR distribution is
a deviatoric stress.